requirements.txt
Flask==2.2.5
Flask-Login==0.6.2
Flask-SQLAlchemy==3.0.3
requests==2.31.0
python-dotenv==1.0.0
stripe
